create login external auth
* adjusted home and login page
* handling response when login successful
* added graphq cliente for get data
* working on redux handle state and actions

Resolve issue #1
