+++ src/pages/Home.jsx
@@ -1,7 +1,14 @@
 import React from "react";
 
-const Home = () => {
-  return <div>you're in the home</div>;
+const Home = ({ user }) => {
+  return (
+    <div>
+      <div id="headerUserInfo">
+        <img src={user?.picture} alt="user img" />
+        <h2>{user?.name}</h2>
+      </div>
+    </div>
+  );
 };
 
 export default Home;
